Removed unused imperative events implementation
