+++ packages/react-native-renderer/src/ReactFabricHostConfig.js
@@ -99,28 +99,6 @@ export type RendererInspectionConfig = $ReadOnly<{
   ) => void,
 }>;
 
-// TODO?: find a better place for this type to live
-export type EventListenerOptions = $ReadOnly<{
-  capture?: boolean,
-  once?: boolean,
-  passive?: boolean,
-  signal: mixed, // not yet implemented
-}>;
-export type EventListenerRemoveOptions = $ReadOnly<{
-  capture?: boolean,
-}>;
-
-// TODO?: this will be changed in the future to be w3c-compatible and allow "EventListener" objects as well as functions.
-export type EventListener = Function;
-
-type InternalEventListeners = {
-  [string]: {
-    listener: EventListener,
-    options: EventListenerOptions,
-    invalidated: boolean,
-  }[],
-};
-
 // TODO: Remove this conditional once all changes have propagated.
 if (registerEventHandler) {
   /**
@@ -137,7 +115,6 @@ class ReactFabricHostComponent {
   viewConfig: ViewConfig;
   currentProps: Props;
   _internalInstanceHandle: Object;
-  _eventListeners: ?InternalEventListeners;
 
   constructor(
     tag: number,
@@ -235,105 +212,6 @@ class ReactFabricHostComponent {
     if (stateNode != null && updatePayload != null) {
       setNativeProps(stateNode.node, updatePayload);
     }
-  }
-
-  // This API (addEventListener, removeEventListener) attempts to adhere to the
-  // w3 Level2 Events spec as much as possible, treating HostComponent as a DOM node.
-  //
-  // Unless otherwise noted, these methods should "just work" and adhere to the W3 specs.
-  // If they deviate in a way that is not explicitly noted here, you've found a bug!
-  //
-  // See:
-  // * https://www.w3.org/TR/DOM-Level-2-Events/events.html
-  // * https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener
-  // * https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/removeEventListener
-  //
-  // And notably, not implemented (yet?):
-  // * https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/dispatchEvent
-  //
-  //
-  // Deviations from spec/TODOs:
-  // (1) listener must currently be a function, we do not support EventListener objects yet.
-  // (2) we do not support the `signal` option / AbortSignal yet
-  addEventListener_unstable(
-    eventType: string,
-    listener: EventListener,
-    options: EventListenerOptions | boolean,
-    // $FlowFixMe[missing-local-annot]
-  ) {
-    if (typeof eventType !== 'string') {
-      throw new Error('addEventListener_unstable eventType must be a string');
-    }
-    if (typeof listener !== 'function') {
-      throw new Error('addEventListener_unstable listener must be a function');
-    }
-
-    // The third argument is either boolean indicating "captures" or an object.
-    const optionsObj =
-      typeof options === 'object' && options !== null ? options : {};
-    const capture =
-      (typeof options === 'boolean' ? options : optionsObj.capture) || false;
-    const once = optionsObj.once || false;
-    const passive = optionsObj.passive || false;
-    const signal = null; // TODO: implement signal/AbortSignal
-
-    /* $FlowFixMe the old version of Flow doesn't have a good way to define an
-     * empty exact object. */
-    const eventListeners: InternalEventListeners = this._eventListeners || {};
-    if (this._eventListeners == null) {
-      this._eventListeners = eventListeners;
-    }
-
-    const namedEventListeners = eventListeners[eventType] || [];
-    if (eventListeners[eventType] == null) {
-      eventListeners[eventType] = namedEventListeners;
-    }
-
-    namedEventListeners.push({
-      listener: listener,
-      invalidated: false,
-      options: {
-        capture: capture,
-        once: once,
-        passive: passive,
-        signal: signal,
-      },
-    });
-  }
-
-  // See https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/removeEventListener
-  removeEventListener_unstable(
-    eventType: string,
-    listener: EventListener,
-    options: EventListenerRemoveOptions | boolean,
-  ) {
-    // eventType and listener must be referentially equal to be removed from the listeners
-    // data structure, but in "options" we only check the `capture` flag, according to spec.
-    // That means if you add the same function as a listener with capture set to true and false,
-    // you must also call removeEventListener twice with capture set to true/false.
-    const optionsObj =
-      typeof options === 'object' && options !== null ? options : {};
-    const capture =
-      (typeof options === 'boolean' ? options : optionsObj.capture) || false;
-
-    // If there are no event listeners or named event listeners, we can bail early - our
-    // job is already done.
-    const eventListeners = this._eventListeners;
-    if (!eventListeners) {
-      return;
-    }
-    const namedEventListeners = eventListeners[eventType];
-    if (!namedEventListeners) {
-      return;
-    }
-
-    // TODO: optimize this path to make remove cheaper
-    eventListeners[eventType] = namedEventListeners.filter(listenerObj => {
-      return !(
-        listenerObj.listener === listener &&
-        listenerObj.options.capture === capture
-      );
-    });
   }
 }
