--- packages/react-native-renderer/src/__mocks__/react-native/Libraries/ReactPrivate/CustomEvent.js
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-'use strict';
-
-// See the react-native repository for a full implementation.
-// This is just a stub, currently to pass `instanceof` checks.
-const CustomEvent = jest.fn();
-
-module.exports = {default: CustomEvent};
