+++ packages/react-native-renderer/src/__mocks__/react-native/Libraries/ReactPrivate/ReactNativePrivateInterface.js
@@ -44,7 +44,4 @@ module.exports = {
   get RawEventEmitter() {
     return require('./RawEventEmitter').default;
   },
-  get CustomEvent() {
-    return require('./CustomEvent').default;
-  },
 };
