--- packages/react-native-renderer/src/legacy-events/PropagationPhases.js
@@ -1,10 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow
- */
-
-export type PropagationPhases = 'bubbled' | 'captured';
